[noise] Remove unneccessary casts
Closes: #9575

+++ network/src/noise/stream.rs
@@ -170,8 +170,7 @@ where
                     decrypted_len,
                     ref mut offset,
                 } => {
-                    let bytes_to_copy =
-                        ::std::cmp::min(decrypted_len as usize - *offset, buf.len());
+                    let bytes_to_copy = ::std::cmp::min(decrypted_len - *offset, buf.len());
                     buf[..bytes_to_copy].copy_from_slice(
                         &self.buffers.read_buffer[*offset..(*offset + bytes_to_copy)],
                     );
@@ -181,7 +180,7 @@ where
                         decrypted_len
                     );
                     *offset += bytes_to_copy;
-                    if *offset == decrypted_len as usize {
+                    if *offset == decrypted_len {
                         self.read_state = ReadState::Init;
                     }
                     return Poll::Ready(Ok(bytes_to_copy));
